requirements.txt
PyYAML==4.2b1
gTTS
wget